admin dashboard and more

+++ src/josh/AllBooks.jsx
@@ -1,8 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 
+
 const AddBook = () => {
+
   const [books, setBooks] = useState([]);
+  
+  const [editBook, setEditBook] = useState(null);
 
   useEffect(() => {
     fetchBooks();
@@ -11,7 +15,7 @@ const AddBook = () => {
   const fetchBooks = () => {
     axios.get("http://localhost:8000/api/getBooks")
       .then((response) => {
-        console.log(response.data);
+        // console.log(response.data);
         setBooks(response.data);
       })
       .catch((error) => {
@@ -19,14 +23,33 @@ const AddBook = () => {
       });
   };
 
-  const handleDeleteBook = (id) => {
-    axios.delete(`http://localhost:8000/api/deleteBooks/${id}`)
-      .then(() => {
+  const handleDeleteBook = (id, booktitle) => {
+
+    const delConf = prompt("Enter book title to confirm book deletion (case-sensitive):")
+
+    if(delConf === booktitle){
+      axios.delete(`http://localhost:8000/api/deleteBooks/${id}`)
+        .then(() => {
         fetchBooks();
       })
       .catch((error) => {
         console.log(error);
       });
+      alert("Book Deletion of "+booktitle+" Successful!")
+    }else{
+      alert("Book Deletion Unsuccessful!")
+    }
+  };
+
+  const handleEditBook = (id) => {
+    axios.put(`http://localhost:8000/api/updateBooks/${id}`, editBook)
+     .then(() => {
+      fetchBooks();
+      setEditBook(null);
+     })
+     .catch((error)=>{
+      console.log(error);
+     });
   };
 
   return (
@@ -60,8 +83,8 @@ const AddBook = () => {
                 <td>{book.shelfno}</td>
                 <td>{book.isbn}</td>
                 <td>
-                  <button>EDIT</button>
-                  <button onClick={() => handleDeleteBook(book._id)}>DELETE</button>
+                  <button onClick={() => setEditBook(book)}>EDIT</button>
+                  <button onClick={() => handleDeleteBook(book._id, book.booktitle)}>DELETE</button>
                 </td>
               </tr>
             ))}
@@ -69,6 +92,99 @@ const AddBook = () => {
         </table>
       </div>
 
+      {editBook && (
+        <>
+        <h1 className='editbooktitle'>EDIT BOOK</h1>
+        <div className='allbookctn'>
+          <div className='allbookctn1'>
+          <label>Book Title</label>
+          <input
+            className='inputbooktitle'
+            type="text"
+            placeholder="Book Title"
+            value={editBook.booktitle}
+            required
+            onChange={(e) => setEditBook({...editBook, booktitle: e.target.value})}
+          />
+          <label>Book Description</label>
+          <textarea
+            className='textareabookdesc'
+            type="text"
+            placeholder="Book Description"
+            value={editBook.bookdesc}
+            required
+            maxLength="150"
+            onChange={(e) => setEditBook({...editBook, bookdesc: e.target.value})}
+          />
+          <label>Book Cover</label>
+          <input
+            type="text"
+            placeholder="Book Cover"
+            value={editBook.bookcover}
+            required
+            onChange={(e) => setEditBook({...editBook, bookcover: e.target.value})}
+          />
+          </div>
+
+          <div className='allbookctn2'>
+          <label>Category</label>
+          <select value={editBook.category} onChange={(e) => setEditBook({...editBook, category: e.target.value})}>
+            <option value="" disabled>Select Category</option>
+            <option>General Knowledge</option>
+            <option>Philosophy and Psychology</option>
+            <option>Religion</option>
+            <option>Social Sciences</option>
+            <option>Languages</option>
+            <option>Science</option>
+            <option>Technology</option>
+            <option>Arts and Recreation</option>
+            <option>Literature</option>
+            <option>History and Geography</option>
+          </select>
+
+          <label>Author</label>
+          <input
+            type="text"
+            placeholder="Author"
+            value={editBook.author}
+            required
+            onChange={(e) => setEditBook({...editBook, author: e.target.value})}
+          />
+          <label>Publisher</label>
+          <input
+            type="text"
+            placeholder="Publisher"
+            value={editBook.publisher}
+            required
+            onChange={(e) => setEditBook({...editBook, publisher: e.target.value})}
+          />
+          <label>Shelf No.</label>
+          <input
+            type="number"
+            min="0"
+            placeholder="Shelf No."
+            value={editBook.shelfno}
+            required
+            onChange={(e) => setEditBook({...editBook, shelfno: e.target.value})}
+          />
+          <label>ISBN</label>
+          <input
+            type="number"
+            min="0"
+            maxLength={13}
+            placeholder="ISBN"
+            value={editBook.isbn}
+            required
+            onChange={(e) => setEditBook({...editBook, isbn: e.target.value})}
+          />
+          </div>
+        </div>
+        <div className='allbookbtnctn'>
+          <button className='editbookbtn' onClick={() => handleEditBook(editBook._id)}>UPDATE</button>
+          <button className='cancelbtn' onClick={() => setEditBook(null)}>CANCEL</button>
+          </div>
+        </>
+      )}
     </>
   );
 };
